moving leadership listing to products
