+++ .github/CODEOWNERS
@@ -19,3 +19,4 @@ src/products/newsStory/ @department-of-veterans-affairs/ap-team
 src/products/storyListing/ @department-of-veterans-affairs/ap-team
 src/products/pressRelease/ @department-of-veterans-affairs/ap-team
 src/products/pressReleaseListing/ @department-of-veterans-affairs/ap-team
+src/products/leadershipListing/ @department-of-veterans-affairs/ap-team @department-of-veterans-affairs/va-platform-cop-frontend
