+++ src/pages/[[...slug]].tsx
@@ -48,7 +48,7 @@ import { VetCenter as FormattedVetCenter } from '@/types/formatted/vetCenter'
 import { HealthCareLocalFacility as FormattedHealthCareLocalFacility } from '@/types/formatted/healthCareLocalFacility'
 import { VamcSystem as FormattedVamcSystem } from '@/types/formatted/vamcSystem'
 import { VamcSystemVaPolice as FormattedVamcSystemVaPolice } from '@/types/formatted/vamcSystemVaPolice'
-import { LeadershipListing as FormattedLeadershipListing } from '@/types/formatted/leadershipListing'
+import { LeadershipListing as FormattedLeadershipListing } from '@/products/leadershipListing/formatted-type'
 // Templates
 import HTMLComment from '@/templates/common/util/HTMLComment'
 import { Event } from '@/products/event/template'
@@ -69,7 +69,7 @@ import { HealthCareLocalFacility } from '@/templates/layouts/healthCareLocalFaci
 import { DoNotPublishError } from '@/lib/drupal/query'
 import { VamcSystem } from '@/templates/layouts/vamcSystem'
 import { VamcSystemVaPolice } from '@/templates/layouts/vamcSystemVaPolice'
-import { LeadershipListing } from '@/templates/layouts/leadershipListing'
+import { LeadershipListing } from '@/products/leadershipListing/template'
 
 // IMPORTANT: in order for a content type to build in Next Build, it must have an appropriate
 // environment variable set in one of two places:
