+++ src/products/leadershipListing/formatted-type.ts
@@ -1,5 +1,5 @@
-import { PublishedEntity } from './publishedEntity'
-import { SideNavMenu } from './sideNav'
+import { PublishedEntity } from '@/types/formatted/publishedEntity'
+import { SideNavMenu } from '@/types/formatted/sideNav'
 import { StaffProfileTeaser } from '@/products/staffProfile/formatted-type'
 import { LovellChildVariant } from '@/lib/drupal/lovell/types'
